+++ src/TcpIpServer.cpp
@@ -84,7 +84,7 @@ int TcpIpServer::Accept()
         return FAIL;
     }
 
-    Socket = BIO_get_fd(_Conn, NULL);
+    Socket = (int)BIO_get_fd(_Conn, NULL);
     TimeVal.tv_sec = _ReadTimeout;
     TimeVal.tv_usec = 0;
     FD_ZERO(&FdSet);
@@ -148,7 +148,7 @@ int TcpIpServer::Read(unsigned char *Buffer, int Length)
         return FAIL;
     }
     
-    Socket = BIO_get_fd(_Client, NULL);
+    Socket = (int)BIO_get_fd(_Client, NULL);
     TimeVal.tv_sec = _ReadTimeout;
     TimeVal.tv_usec = 0;
     FD_ZERO(&FdSet);
